requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
openai==1.99.9
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
SQLAlchemy==2.0.32
greenlet==3.0.3
